Update 'fl' regex and hashbang
Previously, `fl` would match only the first manpage paragraph for a flag
and ignore the last flag listed in a manpage section. The regex now
matches the full flag contents instead and accomodates the last flag in
a section. The hashbang now sets Python3 as the script interpreter
(running it on some remote hosts failed when defaulting to Python2). The
code is itself reformatted to use f-strings rather than C-style print-f
variable specification.