Add shell function 'help'
'help' is a function that is designed as a catch-all for everything
help-related in the shell environment. The query `help <command>`
determines the type of the command (builtin, function, etc) and prints
to the terminal the relevent information related to that command. An
optional `-l` flag queries the linux man-pages rather than the macOS
man-pages.